# Faster optimized frozen dictionary creation (1/n)