-  every strategy needs an array of keys, we can create it up-front and iterate over it rather than the dictionary to get min and max lengths (1-2% gain)
- Instead of ensuring that at least 95% of data is good, we stop when we know that at least 5% is bad (13-14% gain)
- toggle the direction and re-use the comparer and hashset (3% time gain, 12% allocations reduction)

```ini
BenchmarkDotNet=v0.13.2.2052-nightly, OS=Windows 11 (10.0.22621.1702)
AMD Ryzen Threadripper PRO 3945WX 12-Cores, 1 CPU, 24 logical and 12 physical cores
.NET SDK=8.0.100-preview.4.23259.14
  [Host]     : .NET 8.0.0 (8.0.23.25905), X64 RyuJIT AVX2

LaunchCount=3  MaxIterationCount=20  MemoryRandomization=True
```

|                    Method |  Job | Size |      Mean | Ratio | Allocated | Alloc Ratio |
|-------------------------- |----- |----- |----------:|------:|----------:|------------:|
| FrozenDictionaryOptimized |   PR |  512 | 73.236 us |  0.82 |  74.91 KB |        0.88 |
| FrozenDictionaryOptimized | main |  512 | 89.431 us |  1.00 |  85.21 KB |        1.00 |

For this particular benchmark, the initial gap between creating the non-optimized (4 us) and optimized (89.4 us) was 85.4 us, now it's 69.2 us.