@Tornhoof 2023-06-13
> Instead of ensuring that at least 95% of data is good, we stop when we know that at least 5% is bad (13-14% gain)

Do you mean at most 5% is bad? Otherwise you could simply nop everything and 100% is bad, which is at least 5% :)