@stephentoub 2023-06-13
> > Instead of ensuring that at least 95% of data is good, we stop when we know that at least 5% is bad (13-14% gain)
> 
> Do you mean at most 5% is bad? Otherwise you could simply nop everything and 100% is bad, which is at least 5% :)

The wording is correct.  For a given capacity, once we've seen at least 5% of the entries conflict, we know we can't get to > 95% not conflicting, so we can give up on that capacity and try with the next.  We can't give up until we've seen at least 5% conflict, and if we never see 5% conflict, then that capacity is good to go and we can use it.