@IDisposable 2023-06-15
@adamsitnik Can you point me to how to generate the benchmarking analysis? I'm looking into tweaking this a tiny bit more...