@adamsitnik 2023-06-15
> Can you point me to how to generate the benchmarking analysis?

Of course!

[This doc](https://github.com/dotnet/performance/blob/main/docs/benchmarking-workflow-dotnet-runtime.md#dotnet-runtime-prerequisites-for-clr) describes how to run benchmarks against a local build of dotnet/runtime: 

To get a trace file you can use the [EtwProfiler](https://adamsitnik.com/ETW-Profiler/)

Here is a short [introduction to PerfView](https://github.com/dotnet/performance/blob/main/docs/profiling-workflow-dotnet-runtime.md#perfview). 

If you are already familiar with PerfView then the generation of the profile diffs is described [here](https://github.com/dotnet/performance/blob/main/docs/profiling-workflow-dotnet-runtime.md#identifying-regressions)

BTW automating this diff is on my TODO list, I've already made the first step and implemented the https://github.com/dotnet/BenchmarkDotNet/pull/2116

>  I'm looking into tweaking this a tiny bit more...

I am working on it right now, I expect to introduce changes to `FrozenHashTable.Create`  and `CalcNumBuckets` (just letting you know)